requirements.txt
streamlit
requests
streamlit-pdf-viewer
python-dotenv